Several refactors and fixes.
